13th European Conference on Surgery & Cosmetology
Surgery conference will focus on the Advanced Techniques, case studies, Cure, prevention and research in Surgery and cosmetology. Our aim is to bring together bright minds to give talks that are idea-focused, and on a wide range of subjects, to foster learning, inspiration and wonder, and to discuss the novel topics in Surgery.